requirements.txt
pandas
numpy
matplotlib
seaborn
scikit-learn
requests
python-dotenv
jupyter
